+++ packages/svelte/src/internal/client/dev/ownership.js
@@ -1,8 +1,8 @@
 /** @typedef {{ file: string, line: number, column: number }} Location */
 
 import { STATE_SYMBOL } from '../constants.js';
-import { render_effect } from '../reactivity/effects.js';
-import { current_component_context, untrack } from '../runtime.js';
+import { render_effect, user_pre_effect } from '../reactivity/effects.js';
+import { dev_current_component_function } from '../runtime.js';
 import { get_prototype_of } from '../utils.js';
 import * as w from '../warnings.js';
 
@@ -37,7 +37,7 @@ function get_stack() {
  * Determines which `.svelte` component is responsible for a given state change
  * @returns {Function | null}
  */
-function get_component() {
+export function get_component() {
 	// first 4 lines are svelte internals; adjust this number if we change the internal call stack
 	const stack = get_stack()?.slice(4);
 	if (!stack) return null;
@@ -109,8 +109,7 @@ export function mark_module_end(component) {
  */
 export function add_owner(object, owner, global = false) {
 	if (object && !global) {
-		// @ts-expect-error
-		const component = current_component_context.function;
+		const component = dev_current_component_function;
 		const metadata = object[STATE_SYMBOL];
 		if (metadata && !has_owner(metadata, component)) {
 			let original = get_owner(metadata);
@@ -122,6 +121,16 @@ export function add_owner(object, owner, global = false) {
 	}
 
 	add_owner_to_object(object, owner, new Set());
+}
+
+/**
+ * @param {() => unknown} get_object
+ * @param {any} Component
+ */
+export function add_owner_effect(get_object, Component) {
+	user_pre_effect(() => {
+		add_owner(get_object(), Component);
+	});
 }
 
 /**