+++ packages/svelte/src/internal/client/dom/elements/custom-element.js
@@ -193,7 +193,7 @@ if (typeof HTMLElement === 'function') {
 			this.$$cn = false;
 			// In a microtask, because this could be a move within the DOM
 			Promise.resolve().then(() => {
-				if (!this.$$cn) {
+				if (!this.$$cn && this.$$c) {
 					this.$$c.$destroy();
 					destroy_effect(this.$$me);
 					this.$$c = undefined;