+++ packages/svelte/src/internal/client/runtime.js
@@ -115,6 +115,23 @@ export function set_current_component_context(context) {
 	current_component_context = context;
 }
 
+/**
+ * The current component function. Different from current component context:
+ * ```html
+ * <!-- App.svelte -->
+ * <Foo>
+ *   <Bar /> <!-- context == Foo.svelte, function == App.svelte -->
+ * </Foo>
+ * ```
+ * @type {import('#client').ComponentContext['function']}
+ */
+export let dev_current_component_function = null;
+
+/** @param {import('#client').ComponentContext['function']} fn */
+export function set_dev_current_component_function(fn) {
+	dev_current_component_function = fn;
+}
+
 /** @returns {boolean} */
 export function is_runes() {
 	return current_component_context !== null && current_component_context.l === null;
@@ -160,14 +177,17 @@ export function batch_inspect(target, prop, receiver) {
  */
 export function check_dirtiness(reaction) {
 	var flags = reaction.f;
+	var is_dirty = (flags & DIRTY) !== 0;
+	var is_unowned = (flags & UNOWNED) !== 0;
 
-	if ((flags & DIRTY) !== 0) {
+	// If we are unowned, we still need to ensure that we update our version to that
+	// of our dependencies.
+	if (is_dirty && !is_unowned) {
 		return true;
 	}
 
-	if ((flags & MAYBE_DIRTY) !== 0) {
+	if ((flags & MAYBE_DIRTY) !== 0 || (is_dirty && is_unowned)) {
 		var dependencies = reaction.deps;
-		var is_unowned = (flags & UNOWNED) !== 0;
 
 		if (dependencies !== null) {
 			var length = dependencies.length;
@@ -175,26 +195,23 @@ export function check_dirtiness(reaction) {
 			for (var i = 0; i < length; i++) {
 				var dependency = dependencies[i];
 
-				if (check_dirtiness(/** @type {import('#client').Derived} */ (dependency))) {
+				if (!is_dirty && check_dirtiness(/** @type {import('#client').Derived} */ (dependency))) {
 					update_derived(/** @type {import('#client').Derived} **/ (dependency), true);
-
-					// `signal` might now be dirty, as a result of calling `update_derived`
-					if ((reaction.f & DIRTY) !== 0) {
-						return true;
-					}
 				}
 
-				// If we're working with an unowned derived signal, then we need to check
-				// if our dependency write version is higher. If it is then we can assume
-				// that state has changed to a newer version and thus this unowned signal
-				// is also dirty.
-				var version = dependency.version;
-
 				if (is_unowned) {
+					// If we're working with an unowned derived signal, then we need to check
+					// if our dependency write version is higher. If it is then we can assume
+					// that state has changed to a newer version and thus this unowned signal
+					// is also dirty.
+					var version = dependency.version;
+
 					if (version > /** @type {import('#client').Derived} */ (reaction).version) {
 						/** @type {import('#client').Derived} */ (reaction).version = version;
 						return true;
-					} else if (!current_skip_reaction && !dependency?.reactions?.includes(reaction)) {
+					}
+
+					if (!current_skip_reaction && !dependency?.reactions?.includes(reaction)) {
 						// If we are working with an unowned signal as part of an effect (due to !current_skip_reaction)
 						// and the version hasn't changed, we still need to check that this reaction
 						// if linked to the dependency source – otherwise future updates will not be caught.
@@ -205,6 +222,9 @@ export function check_dirtiness(reaction) {
 							reactions.push(reaction);
 						}
 					}
+				} else if ((reaction.f & DIRTY) !== 0) {
+					// `signal` might now be dirty, as a result of calling `check_dirtiness` and/or `update_derived`
+					return true;
 				}
 			}
 		}
@@ -215,7 +235,7 @@ export function check_dirtiness(reaction) {
 		}
 	}
 
-	return false;
+	return is_dirty;
 }
 
 /**
@@ -397,6 +417,11 @@ export function execute_effect(effect) {
 	current_effect = effect;
 	current_component_context = component_context;
 
+	if (DEV) {
+		var previous_component_fn = dev_current_component_function;
+		dev_current_component_function = effect.component_function;
+	}
+
 	try {
 		if ((flags & BLOCK_EFFECT) === 0) {
 			destroy_effect_children(effect);
@@ -408,6 +433,10 @@ export function execute_effect(effect) {
 	} finally {
 		current_effect = previous_effect;
 		current_component_context = previous_component_context;
+
+		if (DEV) {
+			dev_current_component_function = previous_component_fn;
+		}
 	}
 }
 
@@ -880,8 +909,8 @@ export function getContext(key) {
 	const result = /** @type {T} */ (context_map.get(key));
 
 	if (DEV) {
-		// @ts-expect-error
-		const fn = current_component_context.function;
+		const fn = /** @type {import('#client').ComponentContext} */ (current_component_context)
+			.function;
 		if (fn) {
 			add_owner(result, fn, true);
 		}
@@ -935,7 +964,6 @@ export function getAllContexts() {
 	const context_map = get_or_init_context_map('getAllContexts');
 
 	if (DEV) {
-		// @ts-expect-error
 		const fn = current_component_context?.function;
 		if (fn) {
 			for (const value of context_map.values()) {
@@ -983,7 +1011,7 @@ function get_parent_context(component_context) {
  * @returns {number}
  */
 export function update(signal, d = 1) {
-	const value = get(signal);
+	var value = +get(signal);
 	set(signal, value + d);
 	return value;
 }
@@ -994,9 +1022,7 @@ export function update(signal, d = 1) {
  * @returns {number}
  */
 export function update_pre(signal, d = 1) {
-	const value = get(signal) + d;
-	set(signal, value);
-	return value;
+	return set(signal, +get(signal) + d);
 }
 
 /**
@@ -1061,8 +1087,8 @@ export function push(props, runes = false, fn) {
 
 	if (DEV) {
 		// component function
-		// @ts-expect-error
 		current_component_context.function = fn;
+		dev_current_component_function = fn;
 	}
 }
 
@@ -1085,6 +1111,9 @@ export function pop(component) {
 			}
 		}
 		current_component_context = context_stack_item.p;
+		if (DEV) {
+			dev_current_component_function = context_stack_item.p?.function ?? null;
+		}
 		context_stack_item.m = true;
 	}
 	// Micro-optimization: Don't set .a above to the empty object