+++ packages/svelte/tests/runtime-runes/samples/derived-unowned-6/_config.js
@@ -1,0 +1,14 @@
+import { flushSync } from 'svelte';
+import { test } from '../../test';
+
+export default test({
+	async test({ assert, target, logs }) {
+		let [btn1] = target.querySelectorAll('button');
+
+		flushSync(() => {
+			btn1.click();
+		});
+
+		assert.deepEqual(logs, ['computing', 'a', 'a', 'computing', 'bb', 'bb']);
+	}
+});