+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,65 @@
 # svelte
 
+## 5.0.0-next.124
+
+### Patch Changes
+
+- fix: allow to access private fields after `this` reassignment ([#11487](https://github.com/sveltejs/svelte/pull/11487))
+
+- fix: only initiate scroll if scroll binding has existing value ([#11469](https://github.com/sveltejs/svelte/pull/11469))
+
+- fix: restore value after attribute removal during hydration ([#11465](https://github.com/sveltejs/svelte/pull/11465))
+
+- fix: check if svelte component exists on custom element destroy ([#11488](https://github.com/sveltejs/svelte/pull/11488))
+
+- fix: ensure derived is detected as dirty correctly ([#11496](https://github.com/sveltejs/svelte/pull/11496))
+
+- fix: prevent false positive ownership warning ([#11490](https://github.com/sveltejs/svelte/pull/11490))
+
+## 5.0.0-next.123
+
+### Patch Changes
+
+- fix: adjust order of `derived` function definition overloads ([#11426](https://github.com/sveltejs/svelte/pull/11426))
+
+## 5.0.0-next.122
+
+### Patch Changes
+
+- fix: mark function properties on runes as deprecated for better intellisense ([#11439](https://github.com/sveltejs/svelte/pull/11439))
+
+- fix: only warn about non-reactive state in runes mode ([#11434](https://github.com/sveltejs/svelte/pull/11434))
+
+- fix: prevent ownership validation from infering with component context ([#11438](https://github.com/sveltejs/svelte/pull/11438))
+
+- fix: ensure $inspect untracks inspected object ([#11432](https://github.com/sveltejs/svelte/pull/11432))
+
+## 5.0.0-next.121
+
+### Patch Changes
+
+- fix: set correct component context when rendering snippets ([#11401](https://github.com/sveltejs/svelte/pull/11401))
+
+- fix: detect style shorthands as stateful variables in legacy mode ([#11421](https://github.com/sveltejs/svelte/pull/11421))
+
+- fix: improve unowned derived signal behaviour ([#11408](https://github.com/sveltejs/svelte/pull/11408))
+
+- fix: rework binding type-checking strategy ([#11420](https://github.com/sveltejs/svelte/pull/11420))
+
+- fix: improve html escaping of element attributes ([#11411](https://github.com/sveltejs/svelte/pull/11411))
+
+## 5.0.0-next.120
+
+### Patch Changes
+
+- feat: MathML support ([#11387](https://github.com/sveltejs/svelte/pull/11387))
+
+## 5.0.0-next.119
+
+### Patch Changes
+
+- fix: generate correct code for arrow functions with bodies involving object expressions ([#11392](https://github.com/sveltejs/svelte/pull/11392))
+
 ## 5.0.0-next.118
 
 ### Patch Changes